Update README.md
This project is a simple Line Follower Robot that detects and follows a path using IR sensors. The robot reads sensor data and controls motor speeds to stay aligned with the line. It’s built using:

Microcontroller (Arduino/ESP32/etc.)

IR sensors

Motor driver module (L298N)

DC motors and chassis

The logic is based on conditional statements that adjust the direction and speed of the motors according to sensor inputs. The project can be upgraded with PID control for smoother and more accurate movement, as well as additional features like obstacle detection or wireless control.